// src/models/match.ts
export class Match {
    awayteam_country: string;
    awayteam_country_id: number;
    awayteam_id: number;
    awayteam_name: string;
    bookmaker_count: number;
    hometeam_country: string;
    hometeam_country_id: number;
    hometeam_id: number;
    hometeam_name: string;
    id: number;
    last_update: string;
    league_id: number;
    margin: number;
    max_1: number;
    max_2: number;
    max_X: number;
    time: string;
    url: string;
    x_hash: string;
    x_id: number;
    closed: boolean;
    //livescore: any[];
  }